requirements.txt
pywaggle[all]
numpy>=1.20.0
pandas
